refactor: main.js에서 모달을 생성하는 로직 분리
